+++ CHANGELOG.md
@@ -6,11 +6,19 @@
 
 ### Changed
 
-- libs catalog to be merged with hubdle catalog
-
 ### Deprecated
 
 ### Removed
+
+### Fixed
+
+### Updated
+
+## [0.5.0-alpha.11] - 2023-04-18
+
+### Changed
+
+- libs catalog to be merged with hubdle catalog
 
 ### Fixed
 
@@ -217,7 +225,9 @@
 - `org.jetbrains.intellij.plugins:gradle-intellij-plugin -> 1.13.2`
 - `app.cash.molecule:molecule-gradle-plugin -> 0.8.0`
 
-[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.10...HEAD
+[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.11...HEAD
+
+[0.5.0-alpha.11]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.10...0.5.0-alpha.11
 
 [0.5.0-alpha.10]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-alpha.9...0.5.0-alpha.10